+++ server/crates/arachnea-stream/src/services/m6play_resolver.rs
@@ -97,7 +97,12 @@ impl PlayerStreamResolver for M6PlayResolver {
         }
     }
 
-    async fn get_stream(&self, stream_token: &str, body: &[u8]) -> Result<ProxiedStreamResponse> {
+    async fn get_stream(
+        &self,
+        _scraper_agregator: &ScraperAgregator,
+        stream_token: &str,
+        body: &[u8],
+    ) -> Result<ProxiedStreamResponse> {
         proxy_drm_today_license_request(stream_token, body).await
     }
 }