+++ server/crates/arachnea-stream/src/services/rtbf_auvio_resolver.rs
@@ -75,7 +75,12 @@ impl PlayerStreamResolver for RtbfAuvioResolver {
         }
     }
 
-    async fn get_stream(&self, stream_token: &str, body: &[u8]) -> Result<ProxiedStreamResponse> {
+    async fn get_stream(
+        &self,
+        _scraper_agregator: &ScraperAgregator,
+        stream_token: &str,
+        body: &[u8],
+    ) -> Result<ProxiedStreamResponse> {
         proxy_redbee_license_request(stream_token, body).await
     }
 }